:recycle: Refactoring: Member class 구조 변경 (#6)
