+++ src/main/java/forus/naviforyou/global/common/collection/enums/MemberType.java
@@ -1,4 +1,4 @@
-package forus.naviforyou.global.common.entity.enums;
+package forus.naviforyou.global.common.collection.enums;
 
 public enum MemberType {
     KAKAO, NAVER, GOOGLE, GENERAL
