+++ src/main/java/forus/naviforyou/global/common/collection/member/UserAccessibility.java
@@ -1,0 +1,11 @@
+package forus.naviforyou.global.common.collection.member;
+
+
+import forus.naviforyou.global.common.collection.enums.Accessibility;
+
+public class UserAccessibility {
+    private String id; // 건물 id
+    private Accessibility accessibility; // 시설 종류
+}
+
+
